% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_list2df_as_col.R
\name{add_list2df_as_col}
\alias{add_list2df_as_col}
\title{Add List as Column to Dataframe}
\usage{
add_list2df_as_col(dframe, list2add, newcolname = "list2add")
}
\arguments{
\item{dframe}{Dataframe to modify.}

\item{list2add}{A list that has the same length as the number of rows in a dataframe.}

\item{newcolname}{A name for the new column.}
}
\description{
Takes in a dataframe and a list, such that \code{length(list2add) == nrow(dframe)}. 
Each dataframe row is replicated as many times as the length of the corresponding list element. 
After that, the list is unlisted and added to the dataframe.
}
