% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_columns.R
\name{adjust_columns}
\alias{adjust_columns}
\title{Add Columns from Another DataFrame}
\usage{
adjust_columns(df_to_adjust = NULL, df_template = NULL)
}
\arguments{
\item{df_to_adjust}{(dataframe) Dataframe to which new columns will be added.}

\item{df_template}{(dataframe) Dataframe, whose column names will be
compared to \code{df_to_adjust}, and then added to the latter.}
}
\description{
Add columns to one dataframe based on another (template)
such that the first would keep its original ones 
and incorporate all from the second.
}
\author{
Ivan Grishagin
}
