% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_all_matches.R
\name{find_all_matches}
\alias{find_all_matches}
\title{Find All Matches in List}
\usage{
find_all_matches(x_vect, y_vect, separator = "|",
  return_first_match = "FALSE")
}
\arguments{
\item{x_vect}{Vector, elements of which are to be compared.}

\item{y_vect}{Vector, where matches are to be found.}

\item{separator}{See \code{Details}.}

\item{return_first_match}{Logical. Return only the first match? 
In case multiple matches are found, sorts them and returns the first one.}
}
\description{
For each element of one list, finds all matches in the other list.
}
\details{
Returns a vector. By default, merges multiple matches via a \code{separator}.
}
\author{
Ivan Grishagin
}
