% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_in_main.R
\name{funs_in_main}
\alias{funs_in_main}
\title{Functions in Main}
\usage{
funs_in_main(funs_list, str_vect)
}
\description{
First, gets a logical vector of function names that are present in a given vector of strings.\cr
Then for each present function represented as a string vector, 
looks if any other function name is mentioned inside these string vectors
i.e. if the present function relies on some other function on the list.
}
\author{
Ivan Grishagin
}
\keyword{internal}
