% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllFunctionsFromPackage.R
\name{getAllFunctionsFromPackage}
\alias{getAllFunctionsFromPackage}
\title{Extract All Functions from Package}
\usage{
getAllFunctionsFromPackage(package = NULL, path = getwd(),
  one_file = TRUE)
}
\arguments{
\item{package}{Name of the package to be subjected to dismantling.}

\item{path}{Path to directory where to save the file(s) with functions.}

\item{one_file}{(Boolean) One file (\code{TRUE}) or many (\code{FALSE})?}
}
\description{
Extracts all functions from a given package into one or multiple files.
}
\author{
Ivan Grishagin
}
