% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idvect_to_df_by_index.R
\name{idvect_to_df_by_index}
\alias{idvect_to_df_by_index}
\title{Vector to DataFrame}
\usage{
idvect_to_df_by_index(id_vect, split = "-", asnum = TRUE)
}
\arguments{
\item{id_vect}{Vector, elements of which are to be processed.}

\item{split}{Character (not a pattern), on which to split the vector. Defaults to "-".}

\item{asnum}{Logical. Keep only numeric values? Defaults to \code{TRUE}.}
}
\description{
Splits each element of a vector by the \code{split}, and converts the resultant list into a dataframe.
}
\details{
Returns a dataframe. By default, returns only numbers (see \code{asnum}) argument.
}
\author{
Ivan Grishagin
}
