% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackages.R
\name{loadPackages}
\alias{loadPackages}
\title{Install and Load Packages}
\usage{
loadPackages(packages = NULL, popup = TRUE)
}
\arguments{
\item{packages}{Character vector of packages to be loaded/installed.}

\item{popup}{Show a popup notification upon successful package loading?}
}
\description{
Checks which packages are installed, and installs the missing ones from CRAN and BioConductor.
After that, all packages are loaded.
}
\author{
Ivan Grishagin
}
