% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_targz.R
\name{make_targz}
\alias{make_targz}
\title{Make tar.gz Archive}
\usage{
make_targz(input_names, output_name = "default")
}
\arguments{
\item{input_names}{A vector of input file names.}

\item{output_name}{Desired output file name.}
}
\description{
Performs a "tar -cvzf ..." system call. Only works in Linux OS.
}
\author{
Ivan Grishagin
}
