% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_list_in_list.R
\name{match_list_in_list}
\alias{match_list_in_list}
\title{Match List in List}
\usage{
match_list_in_list(list1, list2, ignore = c(NA, "NA", ""),
  only_best_match = TRUE, let_know_when_done = FALSE)
}
\arguments{
\item{list1}{List to compare.}

\item{list2}{List to compare to.}

\item{ignore}{Values to ignore during comparison. Defaults to \code{c(NA,"NA","")}.}

\item{only_best_match}{Return only one best match (i.e. most overlap)? Defaults to TRUE.}

\item{let_know_when_done}{Windows-only! Show a window prompt when done? Defaults to TRUE.}
}
\description{
Compares two lists of vectors. 
Each element of the first list is tested for overlap with all elements of the second list of vectors. 
Indices of the elements, for which such overlap has been found, are returned.
}
\details{
Returns a list that has a length of \code{list1}, and contains \code{list2} indices (or NA values). 
The function is quite efficient: 50K list vs. 4K list with ~50 members per list element comparison takes only ~5 min (i5 laptop).
}
\author{
Ivan Grishagin
}
