% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_fun_files.R
\name{merge_fun_files}
\alias{merge_fun_files}
\title{Merge Files into One File}
\usage{
merge_fun_files(source_dir = NULL, pattern = "\\\\.R",
  result_filename = "merged.R")
}
\arguments{
\item{source_dir}{Path to a directory with files (optional).}

\item{pattern}{(Regular expression) What kind of files to look for?}

\item{result_filename}{Output file name.}
}
\description{
Interactive function that takes in a directory or a list of files 
(using a directory or file picker), and outputs one file with combined contents.
}
\author{
Ivan Grishagin
}
