% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSession.R
\name{prepareSession}
\alias{prepareSession}
\title{Prepare New Session}
\usage{
prepareSession(work_dir = getwd(), create = FALSE)
}
\arguments{
\item{work_dir}{(String) String specifying desired working directory.
If directory does not exist, prompts the user to choose one using a tkwidget (Windows-only).}

\item{create}{(Boolean) Create work directory?}
}
\description{
Function prepares a new session:\cr
     1) sets working directory to the desired one;\cr
     2) sets options(stringsAsFactors = FALSE);\cr
}
\author{
Ivan Grishagin
}
