\name{RIGessentials-package}
\alias{RIGessentials-package}
\alias{RIGessentials}
\docType{package}
\title{
\packageTitle{RIGessentials}
}
\description{
\packageDescription{RIGessentials}
}


\author{
\packageAuthor{RIGessentials}

Maintainer: \packageMaintainer{RIGessentials}
}

\keyword{ package }

\examples{
# set working directory to C:/ drive
prepareSession("C:/")
# install and load packages from CRAN or BioConductor
loadPackages(c("dplyr","limma"))
#strip names from an object
stripNames(list(a=1,b=2))
}
