% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_dependencies.R
\name{source_dependencies}
\alias{source_dependencies}
\title{Source Adjunct R Scripts}
\usage{
source_dependencies(dirs)
}
\arguments{
\item{dirs}{vector of names of directories where to look for the R scripts.}
}
\description{
Source all R scripts in the specified vector of directories.
}
