% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_col_widen_df.R
\name{split_col_widen_df}
\alias{split_col_widen_df}
\title{Split Column Based on Pattern, Expand DF Widthwise}
\usage{
split_col_widen_df(DF, colToSplit, split = "\\\\|", newcolnames = NULL)
}
\arguments{
\item{DF}{Dataframe or data.table object.}

\item{colToSplit}{Column that is to be split.}

\item{split}{Pattern to use for merging of the values. Defaults to pipe (\code{|}).}

\item{newcolnames}{New column names (optional).}
}
\description{
Wrapper for a \code{strsplit} function. 
In a given dataframe, splits values in \code{colToSplit} column using the \col{split} pattern, 
then adds one extra column for each member of a split vector of maximum length to the dataframe. 
If after splitting the number of resultant elements in each row is not the same, fills missing values with NA.
}
\author{
Ivan Grishagin
}
