% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_fun_file_into_many.R
\name{split_fun_file_into_many}
\alias{split_fun_file_into_many}
\title{Split One File with Functions into Many}
\usage{
split_fun_file_into_many(fun_file_path, write_to_files = TRUE)
}
\arguments{
\item{fun_file_path}{Path to file with functions.}

\item{write_to_files}{Write to files? 
If \code{TRUE} (default), will write to files, 
if not -- will return list of functions, 
where each function is represented as a vector of strings.}
}
\description{
Takes in a file with functions and splits it up into separate files.
}
