% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripNames.R
\name{stripNames}
\alias{stripNames}
\title{Strip Names from Object}
\usage{
stripNames(input)
}
\arguments{
\item{input}{Object from which the names are to be removed.}
}
\description{
Takes in an object and tries to remove names.
}
