% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkcheckboxes_from_vect.R
\name{tkcheckboxes_from_vect}
\alias{tkcheckboxes_from_vect}
\title{Tk Checkbox Element From Vector}
\usage{
tkcheckboxes_from_vect(vect, window_title = "Pick one or more options:")
}
\arguments{
\item{vect}{Vector of strings to use as checkbox labels.}

\item{window_title}{Title of the window with radio buttons.}
}
\description{
Takes in a vector and produces a form with TkCheckboxes with those labels as options.
}
\author{
Ivan Grishagin
}
