% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tkradio_from_vect.R
\name{tkradio_from_vect}
\alias{tkradio_from_vect}
\title{Tk Radio Button Element From Vector}
\usage{
tkradio_from_vect(vect, window_title = "Choose an option.")
}
\arguments{
\item{vect}{Vector of strings to use as radio button labels.}

\item{window_title}{Title of the window with radio buttons.}
}
\description{
Takes in a vector and produces a TkRadioButton with those labels as options.
}
