% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_fun_file_based_on_main.R
\name{trim_fun_file_based_on_main}
\alias{trim_fun_file_based_on_main}
\title{Trim Functions File Based on Main}
\usage{
trim_fun_file_based_on_main(fun_file_path, main_file_path,
  one_file = TRUE, save_dir = getwd())
}
\arguments{
\item{fun_file_path}{Path to file with functions.}

\item{main_file_path}{Path to file with main code.}

\item{one_file}{(Boolean) Output as one file? Defaults to \code{TRUE}.}

\item{save_dir}{Where to save the file?}
}
\description{
Takes in a file with functions and a file with "main" code 
then evaluates which functions are present in "main" code
and keeps only those ones.
}
