% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_cols_shorten_df.R
\name{merge_cols_shorten_df}
\alias{merge_cols_shorten_df}
\title{Merge Columns in Dataframe Using Pattern}
\usage{
merge_cols_shorten_df(dFrame = NULL, colKey = NULL, colsToMerge = NULL,
  patternToMerge = "|")
}
\arguments{
\item{dFrame}{Dataframe or data.table object.}

\item{colKey}{Key column. Duplicate values are looked up in this column.}

\item{colsToMerge}{Columns to merge and return. If \code{NULL}, merges all but the \code{colKey} column. 
In \code{colKey} column, duplicate values will be removed, and then this column will be bound to the merged columns.}

\item{patternToMerge}{Pattern to use for merging of the values. Defaults to pipe (\code{|}).}
}
\description{
Looks up duplicate values in the \code{colKey} column, and then merges all corresponding
values in the \code{colToMerge} column, separating them with a \code{patternToMerge} separator.
}
\author{
Ivan Grishagin
}
